read and close response body on http requests
Not reading and closing the response body caused the process to crash
due to too many file descriptors open at the same time. According to
[this blog post][1] the issue is due to the http package opening a file
descriptor for each request, by not closing the body, that file
descriptor effectively remained open forever. The post also mentions
that reading the body until the end improves the performance as well so
even though this program isn't doing anything with the body, I'm adding
a step to read the entire buffer as well.

[1]: https://andrii-kushch.medium.com/is-it-necessary-to-close-the-body-in-the-http-response-object-in-golang-171c44c9394d